backend/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
bcrypt==4.1.3
SQLAlchemy==2.0.31
requests==2.32.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.3.4